Fix more unsigned int / int cases that should be NS[U]Integer
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@31354 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSFont.m
@@ -79,7 +79,7 @@ globalFontMap.
 }
 @end
 @implementation GSFontMapKey
--(unsigned int) hash
+-(NSUInteger) hash
 {
   return hash;
 }
@@ -935,7 +935,7 @@ static void setNSFont(NSString *key, NSFont *font)
   return YES;
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
   int i, sum;
   sum = 0;
